Rework the visualization paragraph and input
