+++ public/components/notebooks/components/input/multi_variant_input.tsx
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React from 'react';
+import React, { useCallback } from 'react';
 import { EuiFlexGroup, EuiInputPopover, EuiSelectable, EuiSpacer } from '@elastic/eui';
 import autosize from 'autosize';
 import { useEffectOnce } from 'react-use';
@@ -17,6 +17,7 @@ import {
   AI_RESPONSE_TYPE,
   DEEP_RESEARCH_PARAGRAPH_TYPE,
 } from '../../../../../common/constants/notebooks';
+import { VisualizationInput } from './visualization_input';
 
 interface MultiVariantInputProps<TParameters = unknown> {
   input?: ParagraphInputType<TParameters>;
@@ -33,6 +34,17 @@ const MultiVariantInputContent: React.FC<{ actionDisabled?: boolean }> = ({ acti
     handleParagraphSelection,
   } = useInputContext();
 
+  const getInputTypeSelector = useCallback(
+    () => (
+      <InputTypeSelector
+        allowSelect
+        current={currInputType}
+        onInputTypeChange={handleSetCurrInputType}
+      />
+    ),
+    [currInputType, handleSetCurrInputType]
+  );
+
   const getInputComponent = () => {
     switch (currInputType) {
       case AI_RESPONSE_TYPE:
@@ -41,24 +53,13 @@ const MultiVariantInputContent: React.FC<{ actionDisabled?: boolean }> = ({ acti
         );
       case 'PPL':
       case 'SQL':
-        return (
-          <QueryPanel
-            prependWidget={
-              <InputTypeSelector
-                allowSelect
-                current={currInputType}
-                onInputTypeChange={handleSetCurrInputType}
-              />
-            }
-            isDisabled={actionDisabled}
-          />
-        );
+        return <QueryPanel prependWidget={getInputTypeSelector()} isDisabled={actionDisabled} />;
       case 'MARKDOWN':
         return <MarkDownInput />;
       case DEEP_RESEARCH_PARAGRAPH_TYPE:
         return <NotebookInput placeholder="Ask question to trigger deep research agent" />;
       case 'VISUALIZATION':
-        return <></>;
+        return <VisualizationInput prependWidget={getInputTypeSelector()} />;
       default:
         return <></>;
     }
@@ -79,15 +80,11 @@ const MultiVariantInputContent: React.FC<{ actionDisabled?: boolean }> = ({ acti
 
   return (
     <>
-      {currInputType !== 'PPL' && currInputType !== 'SQL' && (
+      {currInputType !== 'PPL' && currInputType !== 'SQL' && currInputType !== 'VISUALIZATION' && (
         // Input type selector for query panel is a part of the component already
         <>
           <EuiFlexGroup dir="row" gutterSize="none" justifyContent="spaceBetween">
-            <InputTypeSelector
-              allowSelect
-              current={currInputType}
-              onInputTypeChange={handleSetCurrInputType}
-            />
+            {getInputTypeSelector()}
           </EuiFlexGroup>
           <EuiSpacer size="xs" />
         </>
