+++ public/services/index.ts
@@ -9,11 +9,15 @@ import { DataSourceManagementPluginSetup } from '../../../../src/plugins/data_so
 import { EmbeddableStart } from '../../../../src/plugins/embeddable/public';
 import { ExpressionsStart } from '../../../../src/plugins/expressions/public';
 import { createGetterSetter } from '../../../../src/plugins/opensearch_dashboards_utils/common';
+import { VisualizationsStart } from '../../../../src/plugins/visualizations/public';
 
 export const [getExpressions, setExpressions] = createGetterSetter<ExpressionsStart>('Expressions');
 export const [getData, setData] = createGetterSetter<DataPublicPluginStart>('Data');
 export const [getSearch, setSearch] = createGetterSetter<ISearchStart>('Search');
 export const [getEmbeddable, setEmbeddable] = createGetterSetter<EmbeddableStart>('embeddable');
+export const [getVisualizations, setVisualizations] = createGetterSetter<VisualizationsStart>(
+  'visualizations'
+);
 export const [getDataSourceManagementSetup, setDataSourceManagementSetup] = createGetterSetter<
   | { enabled: true; dataSourceManagement: DataSourceManagementPluginSetup }
   | { enabled: false; dataSourceManagement: undefined }
