+++ public/types.ts
@@ -13,7 +13,10 @@ import {
 import { DataSourceManagementPluginSetup } from '../../../src/plugins/data_source_management/public';
 import { EmbeddableSetup, EmbeddableStart } from '../../../src/plugins/embeddable/public';
 import { NavigationPublicPluginStart } from '../../../src/plugins/navigation/public';
-import { VisualizationsSetup } from '../../../src/plugins/visualizations/public';
+import {
+  VisualizationsSetup,
+  VisualizationsStart,
+} from '../../../src/plugins/visualizations/public';
 import { ExpressionsStart } from '../../../src/plugins/expressions/public';
 import { AppMountParameters, CoreStart } from '../../../src/core/public';
 import PPLService from './services/requests/ppl';
@@ -28,6 +31,7 @@ export interface AppPluginStartDependencies {
   data: DataPublicPluginStart;
   dataSource: DataSourcePluginStart;
   expressions: ExpressionsStart;
+  visualizations: VisualizationsStart;
 }
 
 export interface SetupDependencies {
